Add prompt evaluator module and live mode flag
